+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+v5.7.2
+----------
+ * add status field to index for querying
+
+v5.7.1
+----------
+ * Use contact status instead of is_stopped / is_blocked
+ * Retry HTTP calls to ES
+
+v5.7.0
+----------
+ * Index last_seen_on
+
 v5.6.0
 ----------
  * 5.6.0 Release Candidate